+++ Sorting/radix_sort.cpp
@@ -12,9 +12,9 @@ int getDigitAtIndex(int number, int index) {
 	return digit % 10;
 }
 
-int largestDigitAmount(int array[], size_t length) {
+int largestDigitAmount(int array[], int length) {
 	unsigned int max = 0;
-	for (size_t i = 0; i < length; i++) {
+	for (int i = 0; i < length; i++) {
 		string digits = std::to_string(array[i]);
 		if (max < digits.length()) { max = digits.length(); }
 	}
@@ -23,20 +23,20 @@ int largestDigitAmount(int array[], size_t length) {
 
 void twoDimensionalVectorToArray(vector<std::vector<int>> vec, int arr[]) {
 	int* iterator = arr;
-	for (size_t i = 0; i < vec.size(); i++) {
+	for (int i = 0; i < vec.size(); i++) {
 		std::copy(vec[i].begin(), vec[i].end(), iterator);
 		iterator += vec[i].size();
 	}
 }
 
-void radixSort(int array[], size_t length) {
-	size_t maxDigitAmount = largestDigitAmount(array, length);
-	for (size_t i = 0; i < maxDigitAmount; i++) {
+void radixSort(int array[], int length) {
+	int maxDigitAmount = largestDigitAmount(array, length);
+	for (int i = 0; i < maxDigitAmount; i++) {
 		// 2d vector for storing array items in correct "bucket"
 		vector<vector<int>> digitBuckets;
 		digitBuckets.resize(10);
 
-		for (size_t j = 0; j < length; j++) {
+		for (int j = 0; j < length; j++) {
 			// calculates digit for correct indexing of digitBuckets
 			int digit = getDigitAtIndex(array[j], i);
 			digitBuckets[digit].push_back(array[j]);
@@ -51,15 +51,15 @@ void radixSort(int array[], size_t length) {
 
 
 
-void printArray(int array[], size_t length) {
-	for (size_t i = 0; i < length - 1; i++) {
+void printArray(int array[], int length) {
+	for (int i = 0; i < length - 1; i++) {
 		cout << array[i] << " -- ";
 	}
 	cout << array[length - 1] << endl;
 }
 
 int main() {
-	const size_t arrayLength = 10;
+	const int arrayLength = 10;
 	int array[arrayLength] = {1, 5, 3, 8, 6, 7, 9, 2, 10, 4};
 
 	cout << "INTIAL ARRAY:" << endl;
